Moved models from ingestor to common, refactored ingestor
