+++ apps/ingestor/Dockerfile
@@ -6,6 +6,6 @@ COPY requirements.txt requirements.txt
 
 RUN pip install -r requirements.txt
 
-COPY ./main.py .
+COPY ingestor ingestor
 
-CMD ["uvicorn", "main:app", "--host", "0.0.0.0", "--port", "8000"]
+CMD ["uvicorn", "ingestor.main:app", "--host", "0.0.0.0", "--port", "8000"]
