+++ apps/ingestor/ingestor/main.py
@@ -1,10 +1,16 @@
 import os
-from typing import List, Optional
 
 from dotenv import load_dotenv
 from fastapi import FastAPI
 from pymongo import MongoClient
-from pydantic import BaseModel
+
+from models.raw import (
+    ActorDeathEvent,
+    GameTickEvent,
+    GrandExchangeOfferChangedEvent,
+    HitsplatAppliedEvent,
+    StatChangedEvent,
+)
 
 load_dotenv()
 
@@ -21,95 +27,6 @@ hitsplat_applied_collection = db.get_collection("hitsplat_applied")
 actor_death_collection = db.get_collection("actor_death")
 
 app = FastAPI()
-
-
-class LocationData(BaseModel):
-    x: int
-    y: int
-    regionId: int
-
-
-class ActorData(BaseModel):
-    combatLevel: int
-    location: LocationData
-    name: str
-
-
-class GameTickData(BaseModel):
-    energy: int
-    health: int
-    prayer: int
-    sessionTickCount: int
-    location: LocationData
-    equipmentIds: Optional[List[int]] = None
-
-
-class StatChangedData(BaseModel):
-    boostedLevel: int
-    level: int
-    skill: str
-    xp: int
-
-
-class OfferData(BaseModel):
-    itemId: int
-    price: int
-    quantitySold: int
-    spent: int
-    state: str
-    totalQuantity: int
-
-
-class GrandExchangeOfferData(BaseModel):
-    offer: OfferData
-    slot: int
-
-
-class HitsplatData(BaseModel):
-    amount: int
-    disappearsOnGameCycle: int
-    hitsplatType: int
-    mine: bool
-    others: bool
-
-
-class HitsplatAppliedData(BaseModel):
-    actor: ActorData
-    hitsplat: HitsplatData
-
-
-class ActorDeathData(BaseModel):
-    combatLevel: int
-    location: LocationData
-    name: str
-
-
-class GameEventBase(BaseModel):
-    event: str
-    timestamp: int
-    username: str
-    tickCount: int
-    sessionId: str
-
-
-class GameTickEvent(GameEventBase):
-    data: GameTickData
-
-
-class StatChangedEvent(GameEventBase):
-    data: StatChangedData
-
-
-class GrandExchangeOfferChangedEvent(GameEventBase):
-    data: GrandExchangeOfferData
-
-
-class HitsplatAppliedEvent(GameEventBase):
-    data: HitsplatAppliedData
-
-
-class ActorDeathEvent(GameEventBase):
-    data: ActorDeathData
 
 
 @app.post("/api/v1/event/game-tick/")
